requirements.txt
flask
python-dotenv
requests
pandas
firebase-admin
Flask-Login